third_party/blink/web_tests/external/wpt/css/css-transitions/properties-value-002-expected.txt

This is a testharness.js-based test.
[FAIL] vertical-align vertical(keyword) / values
  assert_not_equals: must not be target value after start got disallowed value "bottom"
[FAIL] vertical-align vertical(keyword) / events
  assert_equals: Expected TransitionEnd events triggered on .transition expected "vertical-align:2s" but got ""
Harness: the test ran to completion.
